refactoring: Move reporter class to cli directory
