+++ lib/src/cli/reporter.dart
